+++ tds/tdstest/surface_completeness_test.go
@@ -28,6 +28,7 @@ func wiredElements() map[string]bool {
 	add("func", functions.RegisteredNames())
 	add("func", exec.EnvScalarNames())
 	add("agg", exec.AggregateNames())
+	add("win", exec.WindowFuncNames())
 	add("sys", sysviews.SupportedViews())
 	add("infoschema", infoschema.SupportedViews())
 	add("proc", engine.SupportedProcs())